Separate ReverseProxy from XML<->JSON conversion
In order to make the Proxy more flexible I changed it from a static
class to an object oriented version. The resulting class creates a
simple one to one ReverseProxy that can change the request and response.
I also added an XML2JSON converting Proxy that inherites the before
mentioned class and converts received XML requests and sends them as
JSON to the server. Responses are converted the other way around.
(Sorry for the long commit & message)

Kim